moved leaflet map state and handlers to a custom hook
